grpc_api/requirements.txt
grpcio
grpcio-tools
grpcio-reflection
sqlalchemy
asyncpg
psycopg2-binary
alembic
python-dotenv
pydantic>=2.0
httpx
